restructure image api to be used by all apps
