+++ docker-compose.dev.yml
@@ -17,7 +17,7 @@ services:
       - .:/code
       # For dev, set the temp folder in the main code directory, for convenience
       - ./_dev_volumes/private_documents:/private_documents
-      - ./_dev_volumes/archive:/archive
+      # - ./_dev_volumes/archive:/archive
       - static_root:/static_map/geocity-wc-map/
     depends_on:
       - "postgres"
