+++ docker-compose.yml
@@ -15,7 +15,7 @@ services:
     command: "gunicorn geocity.wsgi -b :9000 --error-logfile gunicorn_log.log --workers=2 --threads=4 --worker-class=gthread"
     volumes:
       - ${PRIVATE_DOCUMENTS_DIR}:/private_documents
-      - ${ARCHIVE_DIR}:/archive
+      # - ${ARCHIVE_DIR}:/archive
       - static_root:/static_root/geocity-wc-map/
       # to allow to spawn new QGIS containers
       - /var/run/docker.sock:/var/run/docker.sock
