+++ geocity/apps/api/urls.py
@@ -18,4 +18,9 @@ router.register(r"agenda", views.AgendaViewSet, "agenda")
 
 urlpatterns = router.urls + [
     path("current_user/", views.CurrentUserAPIView.as_view(), name="current_user"),
+    path(
+        "image/<int:form_id>/<str:image_name>",
+        views.image_display,
+        name="image_display",
+    ),
 ]
